main,refactor: make error.h main part private
