+++ win32/ctags_vs2013.vcxproj
@@ -256,7 +256,7 @@
     <ClInclude Include="..\main\e_msoft.h" />
     <ClInclude Include="..\main\entry.h" />
     <ClInclude Include="..\main\entry_p.h" />
-    <ClInclude Include="..\main\error.h" />
+    <ClInclude Include="..\main\error_p.h" />
     <ClInclude Include="..\main\field.h" />
     <ClInclude Include="..\main\flags_p.h" />
     <ClInclude Include="..\main\fmt_p.h" />
